+++ tests/dummy/app/controllers/regulatory-statement-sample.ts
@@ -41,10 +41,6 @@ import {
 } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/article-structure-plugin/structures';
 import IntlService from 'ember-intl/services/intl';
 import {
-  variable,
-  variableView,
-} from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/nodes';
-import {
   bullet_list,
   list_item,
   ordered_list,
@@ -66,11 +62,18 @@ import {
   paragraph as paragraphInvisible,
   space,
 } from '@lblod/ember-rdfa-editor/plugins/invisibles';
+import { document_title } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/document-title-plugin/nodes';
 import {
+  codelist,
+  codelistView,
+  location,
+  locationView,
   number,
   numberView,
-} from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/number';
-import { document_title } from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/document-title-plugin/nodes';
+  textVariableView,
+  text_variable,
+} from '@lblod/ember-rdfa-editor-lblod-plugins/plugins/variable-plugin/variables';
+import { VariableConfig } from '@lblod/ember-rdfa-editor-lblod-plugins/components/variable-plugin/insert-variable-card';
 import {
   templateCommentNodes,
   templateCommentView,
@@ -96,7 +99,6 @@ export default class RegulatoryStatementSampleController extends Controller {
       paragraph,
       document_title,
       repaired_block,
-
       list_item,
       ordered_list,
       bullet_list,
@@ -104,12 +106,13 @@ export default class RegulatoryStatementSampleController extends Controller {
       placeholder,
       ...tableNodes({ tableGroup: 'block', cellContent: 'block+' }),
       date: date(this.config.date),
-      variable,
-      number: number,
+      text_variable,
+      number,
+      location,
+      codelist,
       ...STRUCTURE_NODES,
       heading,
       blockquote,
-
       horizontal_rule,
       code_block,
 
@@ -131,6 +134,59 @@ export default class RegulatoryStatementSampleController extends Controller {
       strikethrough,
     },
   });
+
+  get codelistOptions() {
+    return {
+      endpoint: 'https://dev.roadsigns.lblod.info/sparql',
+    };
+  }
+
+  get locationOptions() {
+    return {
+      endpoint: 'https://dev.roadsigns.lblod.info/sparql',
+      zonalLocationCodelistUri:
+        'http://lblod.data.gift/concept-schemes/62331E6900730AE7B99DF7EF',
+      nonZonalLocationCodelistUri:
+        'http://lblod.data.gift/concept-schemes/62331FDD00730AE7B99DF7F2',
+    };
+  }
+
+  get variableTypes(): VariableConfig[] {
+    return [
+      {
+        label: 'text',
+        component: {
+          path: 'variable-plugin/text/insert',
+        },
+      },
+      {
+        label: 'number',
+        component: {
+          path: 'variable-plugin/number/insert',
+        },
+      },
+      {
+        label: 'date',
+        component: {
+          path: 'variable-plugin/date/insert',
+        },
+      },
+      {
+        label: 'location',
+        component: {
+          path: 'variable-plugin/location/insert',
+          options: this.locationOptions,
+        },
+      },
+      {
+        label: 'codelist',
+        component: {
+          path: 'variable-plugin/codelist/insert',
+          options: this.codelistOptions,
+        },
+      },
+    ];
+  }
 
   get config() {
     return {
@@ -165,9 +221,6 @@ export default class RegulatoryStatementSampleController extends Controller {
         ],
         allowCustomFormat: true,
       },
-      variable: {
-        defaultEndpoint: 'https://dev.roadsigns.lblod.info/sparql',
-      },
       templateVariable: {
         endpoint: 'https://dev.roadsigns.lblod.info/sparql',
         zonalLocationCodelistUri:
@@ -190,13 +243,15 @@ export default class RegulatoryStatementSampleController extends Controller {
     controller: SayController,
   ) => Record<string, NodeViewConstructor> = (controller) => {
     return {
-      variable: variableView(controller),
       table_of_contents: tableOfContentsView(this.config.tableOfContents)(
         controller,
       ),
       link: linkView(this.config.link)(controller),
       date: dateView(this.config.date)(controller),
       number: numberView(controller),
+      text_variable: textVariableView(controller),
+      location: locationView(controller),
+      codelist: codelistView(controller),
       templateComment: templateCommentView(controller),
     };
   };